backend/requirements.txt
openai==1.29.0
flask
requests
flask-cors==3.0.9
python-dotenv